Move the `XRefMock` in the unit tests to a central location
This patch helps to avoid code duplication for this mock since more unit
tests are depending on it.
