+++ test/unit/test_utils.js
@@ -14,6 +14,7 @@
  */
 
 import { CMapCompressionType, isNodeJS } from '../../src/shared/util';
+import { isRef } from '../../src/core/primitives';
 
 class NodeFileReaderFactory {
   static fetch(params) {
@@ -74,9 +75,40 @@ class NodeCMapReaderFactory {
   }
 }
 
+class XRefMock {
+  constructor(array) {
+    this._map = Object.create(null);
+
+    for (let key in array) {
+      let obj = array[key];
+      this._map[obj.ref.toString()] = obj.data;
+    }
+  }
+
+  fetch(ref) {
+    return this._map[ref.toString()];
+  }
+
+  fetchAsync(ref) {
+    return Promise.resolve(this.fetch(ref));
+  }
+
+  fetchIfRef(obj) {
+    if (!isRef(obj)) {
+      return obj;
+    }
+    return this.fetch(obj);
+  }
+
+  fetchIfRefAsync(obj) {
+    return Promise.resolve(this.fetchIfRef(obj));
+  }
+}
+
 export {
   NodeFileReaderFactory,
   NodeCMapReaderFactory,
+  XRefMock,
   buildGetDocumentParams,
   TEST_PDFS_PATH,
 };
